+++ client/assets/icons.tsx
@@ -1,0 +1,14 @@
+import * as React from "react"
+
+export const LightningIcon = (props: any) => (
+  <svg
+    width={20}
+    height={20}
+    className="mr-3"
+    viewBox="0 0 200 307"
+    fill="none"
+    {...props}
+  >
+    <path d="m56 0 51.606 131H0l177 176-70.021-142H200L56 0Z" />
+  </svg>
+)
