+++ client/pages/index.tsx
@@ -1,72 +1,30 @@
-import type { NextPage } from 'next'
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
+import { LightningIcon } from '../assets/icons';
+import styles from '../styles/Login.module.css';
 
-const Home: NextPage = () => {
+export default function FirstPost() {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.tsx</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
-  )
+    <section className={styles.container}>
+          <aside className={styles.content}>
+              <h3 className={styles.copy}>Login and we'll automatically create an account 👍🏾</h3>
+        <button className={`${styles.btn} ${styles.btn_primary}`}>
+          <LightningIcon fill="#fff" className={styles.ln_icon} /> Login via
+          Lightning
+        </button>
+        <p>or</p>
+        <form>
+          <label className={styles.input_label} htmlFor="emailInput">Email</label>
+          <input
+            id="emailInput"
+            aria-label="email"
+            className={styles.input}
+            type="email"
+            placeholder="email@example.com"
+          />
+          <button id="email" className={`${styles.btn} ${styles.btn_primary}`}>
+            Login via Email
+          </button>
+        </form>
+      </aside>
+    </section>
+  );
 }
-
-export default Home
